tests/wpt/tests/css/css-ui/parsing/user-find-computed.tentative.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 4: getComputedStyle().userFind</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/3460">
<meta name="assert" content="user-find computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("user-find", "auto");
test_computed_value("user-find", "none");
</script>
</body>
</html>
